added basic logic for determining version
