+++ apps/Standalone/src/App.tsx
@@ -11,8 +11,8 @@ export const App = () => {
     <Routes>
       <Route index element={<DesignerStandalone />} />
       <Route path="/" element={<DesignerStandalone />} />
-      <Route path="datamapperv1" element={<DataMapperV1 />} />
-      <Route path="datamapperv2" element={<DataMapperV2 />} />
+      <Route path="/datamapperv1" element={<DataMapperV1 />} />
+      <Route path="/datamapperv2" element={<DataMapperV2 />} />
       {/* Using path="*"" means "match anything", so this route
                 acts like a catch-all for URLs that we don't have explicit
                 routes for. */}
