+++ apps/vs-code-designer/src/package.json
@@ -762,6 +762,18 @@
               ""
             ]
           },
+          "azureLogicAppsStandard.dataMapperVersion": {
+            "type": "number",
+            "enum": [
+              1,
+              2
+            ],
+            "description": "The default version of Data Mapper to use when performing operations like \"Create new data map\".",
+            "enumDescriptions": [
+              "Version 1",
+              "Version 2"
+            ]
+          },
           "azureLogicAppsStandard.deploySubpath": {
             "scope": "resource",
             "type": "string",
